requirements.txt
certifi==2020.6.20
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1596233721170/work
olefile==0.46
Pillow @ file:///tmp/build/80754af9/pillow_1594307295532/work
six==1.15.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.49.0
